+++ CHANGES.rst
@@ -34,6 +34,9 @@ Mosviz
 Bug Fixes
 ---------
 
+- HST products (as defined by ``OBSTYPE``) are now correctly identified as being
+  either images or 2D spectra. [#4217]
+
 Mosviz
 ^^^^^^